+++ packages/shared/index.ts
@@ -3,14 +3,13 @@ import { Auth } from 'decentraland-auth'
 import './apis/index'
 import './events'
 
-import { ETHEREUM_NETWORK, setNetwork, getTLD, PREVIEW, DEBUG, AVOID_WEB3 } from '../config'
+import { ETHEREUM_NETWORK, setNetwork, getTLD, PREVIEW, DEBUG, ENABLE_WEB3 } from '../config'
 
-import { getUserAccount, getNetwork } from './ethereum/EthereumService'
-import { awaitWeb3Approval } from './ethereum/provider'
 import { initializeUrlPositionObserver } from './world/positionThings'
 import { connect } from './comms'
 import { initialize, queueTrackingEvent } from './analytics'
 import { defaultLogger } from './logger'
+import { initWeb3, getNetworkFromTLD } from './web3'
 
 // TODO fill with segment keys and integrate identity server
 export async function initializeAnalytics(userId: string) {
@@ -25,43 +24,6 @@ export async function initializeAnalytics(userId: string) {
     default:
       return initialize('a4h4BC4dL1v7FhIQKKuPHEdZIiNRDVhc', userId)
   }
-}
-
-function getNetworkFromTLD(): ETHEREUM_NETWORK | null {
-  const tld = getTLD()
-  if (tld === 'zone') {
-    return ETHEREUM_NETWORK.ROPSTEN
-  }
-
-  if (tld === 'today' || tld === 'org') {
-    return ETHEREUM_NETWORK.MAINNET
-  }
-
-  return null
-}
-
-async function getAddress(): Promise<string | undefined> {
-  try {
-    await awaitWeb3Approval()
-    return await getUserAccount()
-  } catch (e) {
-    defaultLogger.info(e)
-  }
-}
-
-async function getAppNetwork(): Promise<ETHEREUM_NETWORK> {
-  const web3Network = await getNetwork()
-  const web3net = web3Network === '1' ? ETHEREUM_NETWORK.MAINNET : ETHEREUM_NETWORK.ROPSTEN
-  // TLD environment have priority
-  const net = getNetworkFromTLD() || web3net
-
-  if (web3net && net !== web3net) {
-    // TODO @fmiras show an HTML error if web3 networks differs from domain network and do not load client at all
-    defaultLogger.error(`Switch to network ${net}`)
-  }
-
-  defaultLogger.info('Using ETH network: ', net)
-  return net
 }
 
 export async function initShared(container: HTMLElement): Promise<ETHEREUM_NETWORK> {
@@ -93,14 +55,12 @@ export async function initShared(container: HTMLElement): Promise<ETHEREUM_NETWO
   console['groupEnd']()
 
   console['group']('connect#ethereum')
-  const address = await getAddress()
 
-  if (address) {
-    defaultLogger.log(`Identifying address ${address}`)
-    queueTrackingEvent('Use web3 address', { address })
+  if (ENABLE_WEB3) {
+    await initWeb3()
   }
 
-  const net = await getAppNetwork()
+  const net = (await getNetworkFromTLD()) || ETHEREUM_NETWORK.MAINNET
   queueTrackingEvent('Use network', { net })
 
   // Load contracts from https://contracts.decentraland.org
@@ -111,15 +71,14 @@ export async function initShared(container: HTMLElement): Promise<ETHEREUM_NETWO
   await connect(
     user_id,
     net,
-    auth,
-    address
+    auth
   )
   console['groupEnd']()
 
   initializeUrlPositionObserver()
 
   // Warn in case wallet is set in mainnet
-  if (net === ETHEREUM_NETWORK.MAINNET && DEBUG && !AVOID_WEB3) {
+  if (net === ETHEREUM_NETWORK.MAINNET && DEBUG && ENABLE_WEB3) {
     const style = document.createElement('style')
     style.appendChild(
       document.createTextNode(
